refactor: conversion to hilt from dagger2 success
