+++ build.gradle
@@ -8,6 +8,7 @@ plugins {
     id 'com.android.library' version '7.2.0' apply false
     id 'org.jetbrains.kotlin.android' version '1.5.32' apply false
     id 'org.jetbrains.kotlin.jvm' version '1.5.32' apply false
+    id 'com.google.dagger.hilt.android' version "2.41" apply false
 }
 
 tasks.register("clean", Delete) {
