+++ data/src/main/java/com/haomins/data/di/ConnectionModule.kt
@@ -1,21 +1,23 @@
-package com.haomins.data.di.module
+package com.haomins.data.di
 
 import com.haomins.data.service.TheOldReaderService
 import com.jakewharton.retrofit2.adapter.rxjava2.RxJava2CallAdapterFactory
 import dagger.Module
 import dagger.Provides
-import dagger.Reusable
+import dagger.hilt.InstallIn
+import dagger.hilt.components.SingletonComponent
 import io.reactivex.schedulers.Schedulers
 import okhttp3.OkHttpClient
 import okhttp3.logging.HttpLoggingInterceptor
 import retrofit2.Retrofit
 import retrofit2.converter.gson.GsonConverterFactory
+import javax.inject.Singleton
 
 @Module
+@InstallIn(SingletonComponent::class)
 object ConnectionModule {
 
-    @Reusable
-    @JvmStatic
+    @Singleton
     @Provides
     fun provideTheOldReaderService(okHttpClient: OkHttpClient): TheOldReaderService {
         return Retrofit.Builder().apply {
@@ -26,8 +28,7 @@ object ConnectionModule {
         }.build().create(TheOldReaderService::class.java)
     }
 
-    @Reusable
-    @JvmStatic
+    @Singleton
     @Provides
     fun provideOkHttpClient(): OkHttpClient {
         val httpLogInterceptor = HttpLoggingInterceptor()
