--- data/src/main/java/com/haomins/data/di/DataModule.kt
@@ -1,13 +1,0 @@
-package com.haomins.data.di
-
-import com.haomins.data.di.module.*
-import dagger.Module
-
-@Module(includes = [
-    ConnectionModule::class,
-    PreferenceModule::class,
-    RepositoryModule::class,
-    RxModule::class,
-    SchedulerModule::class
-])
-interface DataModule
