+++ data/src/main/java/com/haomins/data/di/PreferenceModule.kt
@@ -1,19 +1,21 @@
-package com.haomins.data.di.module
+package com.haomins.data.di
 
 import android.app.Application
 import android.content.Context
 import android.content.SharedPreferences
 import dagger.Module
 import dagger.Provides
-import dagger.Reusable
+import dagger.hilt.InstallIn
+import dagger.hilt.components.SingletonComponent
+import javax.inject.Singleton
 
 @Module
+@InstallIn(SingletonComponent::class)
 object PreferenceModule {
 
     private const val SHARED_PREFERENCE_NAME = "READER_SHARED_PREF"
 
-    @Reusable
-    @JvmStatic
+    @Singleton
     @Provides
     fun provideSharedPreference(application: Application): SharedPreferences {
         return application.getSharedPreferences(
