+++ data/src/main/java/com/haomins/data/di/RepositoryModule.kt
@@ -1,11 +1,14 @@
-package com.haomins.data.di.module
+package com.haomins.data.di
 
 import com.haomins.data.repositories.*
 import com.haomins.domain.repositories.*
 import dagger.Binds
 import dagger.Module
+import dagger.hilt.InstallIn
+import dagger.hilt.android.components.ViewModelComponent
 
 @Module
+@InstallIn(ViewModelComponent::class)
 interface RepositoryModule {
 
     @Binds
