+++ presentation/build.gradle
@@ -3,6 +3,7 @@ plugins {
     id "kotlin-android"
     id "kotlin-android-extensions"
     id "kotlin-kapt"
+    id 'com.google.dagger.hilt.android'
 }
 
 android {
