--- presentation/src/main/java/com/haomins/reader/di/modules/PresentationModule.kt
@@ -1,7 +1,0 @@
-package com.haomins.reader.di.modules
-
-import com.haomins.reader.viewModels.di.ViewModelComponent
-import dagger.Module
-
-@Module(subcomponents = [ViewModelComponent::class])
-object PresentationModule
