+++ presentation/src/main/java/com/haomins/reader/view/fragments/AddSourceFragment.kt
@@ -9,24 +9,22 @@ import androidx.core.widget.doOnTextChanged
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.viewModels
 import androidx.lifecycle.Observer
-import androidx.lifecycle.ViewModelProvider
 import com.haomins.reader.R
 import com.haomins.reader.ReaderApplication
 import com.haomins.reader.utils.hideKeyboard
 import com.haomins.reader.utils.showSnackbar
 import com.haomins.reader.viewModels.AddSourceViewModel
+import dagger.hilt.android.AndroidEntryPoint
 import kotlinx.android.synthetic.main.fragment_add_source.*
-import javax.inject.Inject
 
+@AndroidEntryPoint
 class AddSourceFragment : Fragment() {
 
     companion object {
         const val TAG = "AddSourceFragment"
     }
 
-    @Inject
-    lateinit var viewModelFactory: ViewModelProvider.Factory
-    private val addSourceViewModel by viewModels<AddSourceViewModel> { viewModelFactory }
+    private val addSourceViewModel by viewModels<AddSourceViewModel>()
 
     private val isSourceAddedObserver by lazy {
         Observer<Pair<AddSourceViewModel.AddSourceStatus, String>> {
@@ -43,11 +41,6 @@ class AddSourceFragment : Fragment() {
                 AddSourceViewModel.AddSourceStatus.DEFAULT -> Unit
             }
         }
-    }
-
-    override fun onAttach(context: Context) {
-        (requireActivity().application as ReaderApplication).appComponent.viewModelComponent().build().inject(this)
-        super.onAttach(context)
     }
 
     override fun onCreateView(
