+++ presentation/src/main/java/com/haomins/reader/view/fragments/SettingsFragment.kt
@@ -12,8 +12,10 @@ import com.haomins.reader.R
 import com.haomins.reader.ReaderApplication
 import com.haomins.reader.view.activities.SettingsActivity
 import com.haomins.reader.viewModels.SettingsViewModel
+import dagger.hilt.android.AndroidEntryPoint
 import javax.inject.Inject
 
+@AndroidEntryPoint
 class SettingsFragment : PreferenceFragmentCompat() {
 
     companion object {
@@ -25,18 +27,10 @@ class SettingsFragment : PreferenceFragmentCompat() {
         private const val INTENT_EMAIL_TYPE = "message/rfc822"
     }
 
-    @Inject
-    lateinit var viewModelFactory: ViewModelProvider.Factory
-    private val settingsViewModel by viewModels<SettingsViewModel> { viewModelFactory }
+    private val settingsViewModel by viewModels<SettingsViewModel>()
 
     private val darkModeSwitchPreferenceCompat by lazy {
         findPreference<SwitchPreferenceCompat>(DARK_MODE_OPTION)
-    }
-
-    override fun onAttach(context: Context) {
-        (requireActivity().application as ReaderApplication).appComponent.viewModelComponent()
-            .build().inject(this)
-        super.onAttach(context)
     }
 
     override fun onCreatePreferences(savedInstanceState: Bundle?, rootKey: String?) {
