+++ presentation/src/main/java/com/haomins/reader/viewModels/SettingsViewModel.kt
@@ -10,7 +10,9 @@ import com.haomins.reader.utils.DarkModeManager
 import io.reactivex.observers.DisposableSingleObserver
 import java.io.File
 import javax.inject.Inject
+import dagger.hilt.android.lifecycle.HiltViewModel
 
+@HiltViewModel
 class SettingsViewModel @Inject constructor(
     private val darkModeManager: DarkModeManager,
     private val getLogFiles: GetLogFiles,
