--- presentation/src/main/java/com/haomins/reader/viewModels/di/CustomViewModelFactory.kt
@@ -1,32 +1,0 @@
-package com.haomins.reader.viewModels.di
-
-import androidx.lifecycle.ViewModel
-import androidx.lifecycle.ViewModelProvider
-
-import javax.inject.Inject
-import javax.inject.Provider
-
-class CustomViewModelFactory @Inject constructor(
-        private val creators: @JvmSuppressWildcards Map<Class<out ViewModel>, Provider<ViewModel>>
-) : ViewModelProvider.Factory {
-    override fun <T : ViewModel> create(modelClass: Class<T>): T {
-        var creator: Provider<out ViewModel>? = creators[modelClass]
-        if (creator == null) {
-            for ((key, value) in creators) {
-                if (modelClass.isAssignableFrom(key)) {
-                    creator = value
-                    break
-                }
-            }
-        }
-        if (creator == null) {
-            throw IllegalArgumentException("Unknown model class: $modelClass")
-        }
-        try {
-            @Suppress("UNCHECKED_CAST")
-            return creator.get() as T
-        } catch (e: Exception) {
-            throw RuntimeException(e)
-        }
-    }
-}
