--- presentation/src/main/java/com/haomins/reader/viewModels/di/ViewModelComponent.kt
@@ -1,34 +1,0 @@
-package com.haomins.reader.viewModels.di
-
-import com.haomins.reader.view.activities.MainActivity
-import com.haomins.reader.view.fragments.*
-import com.haomins.reader.viewModels.di.modules.ViewModelModule
-import dagger.Subcomponent
-
-@Subcomponent(modules = [ViewModelModule::class])
-interface ViewModelComponent {
-
-    // fragments
-    fun inject(addSourceFragment: AddSourceFragment)
-
-    fun inject(articleDetailFragment: ArticleDetailFragment)
-
-    fun inject(articleListFragment: ArticleListFragment)
-
-    fun inject(loginFragment: LoginFragment)
-
-    fun inject(settingsFragment: SettingsFragment)
-
-    fun inject(sourceTitleListFragment: SourceTitleListFragment)
-
-    fun inject(disclosureFragment: DisclosureFragment)
-
-    // activities
-    fun inject(mainActivity: MainActivity)
-
-    @Subcomponent.Builder
-    interface Builder {
-        fun build(): ViewModelComponent
-    }
-
-}
