--- presentation/src/main/java/com/haomins/reader/viewModels/di/ViewModelKey.kt
@@ -1,15 +1,0 @@
-package com.haomins.reader.viewModels.di
-
-import androidx.lifecycle.ViewModel
-import dagger.MapKey
-import kotlin.reflect.KClass
-
-@MustBeDocumented
-@Target(
-        AnnotationTarget.FUNCTION,
-        AnnotationTarget.PROPERTY_GETTER,
-        AnnotationTarget.PROPERTY_SETTER
-)
-@Retention(AnnotationRetention.RUNTIME)
-@MapKey
-annotation class ViewModelKey(val value: KClass<out ViewModel>)
